Initialization didn't work if the static folder already existed (when not default location)

+++ blogapp/views.py
@@ -233,7 +233,8 @@ def initialize():
     db.session.commit()
 
     # Copy of the useful static file for application
-    if current_app.config['STATIC_FOLDER'] != 'blogapp/static':
+    if (current_app.config['STATIC_FOLDER'] != 'blogapp/static'
+        and not os.path.exists(current_app.config['STATIC_FOLDER'])):
         shutil.copytree('blogapp/static', current_app.config['STATIC_FOLDER'])
 
     # Download of the external static file
